added 'entiy' signifier to public domain entity, moved into it's own module
